Retrieval/requirements.txt
matplotlib==3.7.2
numpy==1.25.2
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1

